refactor: se movieron todos los estilos a un archivo externo
